agregamos interface Direccion y más test

+++ src/main/kotlin/ar/edu/unahur/obj2/socios/Animo.kt
@@ -2,25 +2,25 @@ package ar.edu.unahur.obj2.socios
 
 
 interface Animo {
-    fun afectarPropinaSegunAnimo(cliente: Cliente, pedido: Pedido) : Float
+    fun afectarPropinaSegunAnimo(cliente: Cliente, pedido: Pedido) : Double
 }
 object Enojado : Animo{
-    override fun afectarPropinaSegunAnimo(cliente: Cliente, pedido: Pedido): Float {
-        return 0F
+    override fun afectarPropinaSegunAnimo(cliente: Cliente, pedido: Pedido): Double {
+        return 0.0
     }
 }
 object Feliz : Animo {
-    override fun afectarPropinaSegunAnimo(cliente: Cliente, pedido: Pedido): Float {
-        return (pedido.valor * 0.25).toFloat()
+    override fun afectarPropinaSegunAnimo(cliente: Cliente, pedido: Pedido): Double {
+        return (pedido.valor * 0.25)
     }
 }
 object Indiferente : Animo{
-    override fun afectarPropinaSegunAnimo(cliente: Cliente, pedido: Pedido): Float {
+    override fun afectarPropinaSegunAnimo(cliente: Cliente, pedido: Pedido): Double {
         return cliente.plataBolsillo
     }
 }
 object Resfriado : Animo{
-    override fun afectarPropinaSegunAnimo(cliente: Cliente, pedido: Pedido): Float {
+    override fun afectarPropinaSegunAnimo(cliente: Cliente, pedido: Pedido): Double {
         return pedido.valor
     }
 }
